refactor: split utilities and relocate global styles
