+++ components.json
@@ -5,7 +5,7 @@
   "tsx": true,
   "tailwind": {
     "config": "",
-    "css": "src/app/globals.css",
+    "css": "src/shared/utils/theme.css",
     "baseColor": "neutral",
     "cssVariables": true,
     "prefix": ""
