+++ src/shared/ui/Spinner.tsx
@@ -2,7 +2,7 @@
 'use client';
 
 import React from 'react';
-import { cn } from '@/shared/utils/utils';
+import { cn } from '@/shared/utils/cn';
 
 interface SpinnerProps {
   className?: string;
