+++ src/shared/ui/calendar.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import { ChevronDownIcon, ChevronLeftIcon, ChevronRightIcon } from 'lucide-react';
 import { DayButton, DayPicker, getDefaultClassNames } from 'react-day-picker';
 
-import { cn } from '@/shared/utils/utils';
+import { cn } from '@/shared/utils/cn';
 import { Button, buttonVariants } from '@/shared/ui/button';
 
 function Calendar({
