+++ src/shared/ui/popover.tsx
@@ -4,7 +4,7 @@
 import * as React from 'react';
 import * as PopoverPrimitive from '@radix-ui/react-popover';
 
-import { cn } from '@/shared/utils/utils';
+import { cn } from '@/shared/utils/cn';
 
 function Popover({ ...props }: React.ComponentProps<typeof PopoverPrimitive.Root>) {
   return <PopoverPrimitive.Root data-slot="popover" {...props} />;
