+++ src/shared/utils/capitalize.ts
@@ -1,0 +1,8 @@
+// src/shared/utils/capitalize.ts
+
+/**
+ * Capitalizes the first letter of a string.
+ */
+export function capitalize(str: string) {
+  return str ? str.charAt(0).toUpperCase() + str.slice(1) : '';
+}
