+++ src/shared/utils/cn.ts
@@ -1,0 +1,8 @@
+// src/shared/utils/cn.ts
+
+import { clsx, type ClassValue } from 'clsx';
+import { twMerge } from 'tailwind-merge';
+
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
