--- src/shared/utils/utils.ts
@@ -1,17 +1,0 @@
-// src/shared/utils/utils.ts
-
-import { clsx, type ClassValue } from 'clsx';
-import { twMerge } from 'tailwind-merge';
-
-export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs));
-}
-
-/**
- * Capitalizes the first letter of a string.
- */
-export function capitalize(str: string) {
-  return str ? str.charAt(0).toUpperCase() + str.slice(1) : '';
-}
-
-// Note: Removed measureTextWidth in favor of inlining logic where needed.
